Code cleanup w/ Nasih
